Use marquee::marquee_parse() to format Markdown text

fmt_marquee_txt() uses marquee::marquee_parse() and openxlsx2::fmt_txt() to format Markdown text.

Usage

fmt_marquee_txt(
  text,
  ...,
  style = marquee::classic_style(),
  ignore_html = TRUE
)

Arguments

text

A character string. The core quality of markdown is that any text is valid markdown so there is no restrictions on the content

...

Ignored at this time. For future use.

style

A style set such as classic_style() that defines how the text should be rendered

ignore_html

Should HTML code be removed from the output

Examples

fmt_marquee_txt(
  "# ABC

  abc"
)
#> Error in purrr::pmap(parsed_text, function(text, type, indentation, indent,     size, ol_index, weight, bullets, color, underline, strikethrough,     ...) {    before <- paste0(rep(" ", indent), collapse = "")    if (type == "li") {        bullet <- bullets[[indentation - 2]]        before <- paste0(before, bullet, " ")    }    else if (ol_index > 0) {        before <- paste0(before, ol_index, ". ")    }    openxlsx2::fmt_txt(x = paste0(before, text), bold = weight >         400, underline = underline, strike = strikethrough, size = size,         color = color)}): ℹ In index: 1.
#> Caused by error:
#> ! color must be class wbColour