Convert an object to a list of workbooks

map_wb() takes a list and returns a list of wbWorkbook objects.

Usage

map_wb(x, ...)

Arguments

x

Typically, a data frame or list of data frames. A wbWorkbook is returned as-is ignoring all other parameters. If type = "any", x can also be an object that is coercible to a data frame.

...

Arguments passed on to as_wb

type

Type of objects to allow. "df-list" allows data frames and lists of data frames. "df" allows data frames only. "any" allows any input (allowing the option for wb_add_data_ext() to coerce objects to data frames).

sheet_names

Optional character vector of worksheet names.

title,subject,category,keywords

Additional arguments passed to openxlsx2::wb_workbook().

properties

A named list (typically from openxlsx2::wb_get_properties()) used to set new workbook properties for any values set to NULL. datetime_created defaults to Sys.time() so must be set to NULL to inherit value from properties.

creator

Creator of the workbook (your name). Defaults to login username or options("openxlsx2.creator") if set.

datetime_created

The time of the workbook is created

theme

Optional theme identified by string or number. See Details for options.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

A list of wbWorkbook objects.