Use dplyr::group_split to split a workbook into a list of workbooks

wb_split() uses wb_to_df_list() to extract the data frames from each sheet of a workbook and then split the

Usage

wb_split(wb, ..., .key = NULL)

Arguments

...

Arguments passed on to wb_to_df_list

sheet_names

Character vector of sheet names. If not supplied, all sheet names from the supplied workbook are used.

.key

Passed to dplyr::group_split().

Value

A list of wbWorkbook objects.

Examples

wb <- as_wb(list(mtcars, mtcars))

wb_split(wb, .key = carb)
#> Warning: unused arguments (wb)
#> Error in (function (file, sheet, start_row = 1, start_col = NULL, row_names = FALSE,     col_names = TRUE, skip_empty_rows = FALSE, skip_empty_cols = FALSE,     skip_hidden_rows = FALSE, skip_hidden_cols = FALSE, rows = NULL,     cols = NULL, detect_dates = TRUE, na.strings = "#N/A", na.numbers = NA,     fill_merged_cells = FALSE, dims, show_formula = FALSE, convert = TRUE,     types, named_region, keep_attributes = FALSE, check_names = FALSE,     show_hyperlinks = FALSE, ...) {    xlsx_file <- list(...)$xlsx_file    standardize_case_names(...)    if (!is.null(xlsx_file)) {        .Deprecated(old = "xlsx_file", new = "file", package = "openxlsx2")        file <- xlsx_file %||% file    }    if (!is.null(cols))         cols <- col2int(cols)    if (inherits(file, "wbWorkbook")) {        wb <- file    }    else {        if (missing(sheet))             sheet <- substitute()        data_only <- TRUE        if (!missing(named_region) || show_hyperlinks)             data_only <- FALSE        wb <- wb_load(file, sheet = sheet, data_only = data_only)    }    if (!missing(named_region)) {        nr <- wb$get_named_regions(tables = TRUE)        if ((named_region %in% nr$name) && missing(sheet)) {            sel <- nr[nr$name == named_region, ][1, ]            sheet <- sel$sheet            dims <- sel$coords        }        else if (named_region %in% nr$name) {            sel <- nr[nr$name == named_region & nr$sheet == wb_validate_sheet(wb,                 sheet), ]            if (NROW(sel) == 0) {                stop("no such named_region on selected sheet")            }            dims <- sel$coords        }        else {            stop("no such named_region")        }    }    if (missing(sheet)) {        sheet <- 1    }    if (is.character(sheet)) {        sheet <- wb_validate_sheet(wb, sheet)    }    if (is.null(wb$worksheets[[sheet]]$sheet_data$cc) || nrow(wb$worksheets[[sheet]]$sheet_data$cc) ==         0) {        message("sheet found, but contains no data")        return(NULL)    }    if (missing(named_region) && missing(dims)) {        sd <- wb$worksheets[[sheet]]$sheet_data$cc[c("row_r",             "c_r")]        sd$row <- as.integer(sd$row_r)        sd$col <- col2int(sd$c_r)        dims <- paste0(int2col(min(sd$col)), min(sd$row), ":",             int2col(max(sd$col)), max(sd$row))    }    row_attr <- wb$worksheets[[sheet]]$sheet_data$row_attr    cc <- wb$worksheets[[sheet]]$sheet_data$cc    sst <- attr(wb$sharedStrings, "text")    rnams <- row_attr$r    numfmt_date <- numfmt_is_date(wb$styles_mgr$styles$numFmts)    xlsx_date_style <- style_is_date(wb$styles_mgr$styles$cellXfs,         numfmt_date)    numfmt_hms <- numfmt_is_hms(wb$styles_mgr$styles$numFmts)    xlsx_hms_style <- style_is_hms(wb$styles_mgr$styles$cellXfs,         numfmt_hms)    numfmt_posix <- numfmt_is_posix(wb$styles_mgr$styles$numFmts)    xlsx_posix_style <- style_is_posix(wb$styles_mgr$styles$cellXfs,         numfmt_posix)    z <- dims_to_dataframe(dims, empty_rm = TRUE)    tt <- copy(z)    keep_cols <- colnames(z)    keep_rows <- rownames(z)    maxRow <- max(as.numeric(keep_rows))    maxCol <- max(col2int(keep_cols))    if (start_row > 1) {        keep_rows <- as.character(seq(start_row, maxRow))        if (start_row <= maxRow) {            z <- z[rownames(z) %in% keep_rows, , drop = FALSE]            tt <- tt[rownames(tt) %in% keep_rows, , drop = FALSE]        }        else {            keep_rows <- as.character(start_row)            z <- z[keep_rows, , drop = FALSE]            tt <- tt[keep_rows, , drop = FALSE]            rownames(z) <- as.integer(keep_rows)            rownames(tt) <- as.integer(keep_rows)        }    }    if (!is.null(rows)) {        keep_rows <- as.character(as.integer(rows))        if (all(keep_rows %in% rownames(z))) {            z <- z[rownames(z) %in% keep_rows, , drop = FALSE]            tt <- tt[rownames(tt) %in% keep_rows, , drop = FALSE]        }        else {            z <- z[keep_rows, , drop = FALSE]            tt <- tt[keep_rows, , drop = FALSE]            rownames(z) <- as.integer(keep_rows)            rownames(tt) <- as.integer(keep_rows)        }    }    if (!is.null(start_col)) {        keep_cols <- int2col(seq(col2int(start_col), maxCol))        if (!all(keep_cols %in% colnames(z))) {            keep_col <- keep_cols[!keep_cols %in% colnames(z)]            z[keep_col] <- NA_character_            tt[keep_col] <- NA_character_            z <- z[keep_cols]            tt <- tt[keep_cols]        }        z <- z[, match(keep_cols, colnames(z)), drop = FALSE]        tt <- tt[, match(keep_cols, colnames(tt)), drop = FALSE]    }    if (!is.null(cols)) {        keep_cols <- int2col(cols)        if (!all(keep_cols %in% colnames(z))) {            keep_col <- keep_cols[!keep_cols %in% colnames(z)]            z[keep_col] <- NA_character_            tt[keep_col] <- NA_character_        }        z <- z[, match(keep_cols, colnames(z)), drop = FALSE]        tt <- tt[, match(keep_cols, colnames(tt)), drop = FALSE]    }    keep_rows <- keep_rows[keep_rows %in% rnams]    if (length(keep_rows) && length(keep_cols))         cc <- cc[cc$row_r %in% keep_rows & cc$c_r %in% keep_cols,             ]    cc$val <- NA_character_    cc$typ <- NA_character_    cc_tab <- unique(cc$c_t)    if (any(cc_tab == "b")) {        sel <- cc$c_t %in% "b"        cc$val[sel] <- as.logical(as.numeric(cc$v[sel]))        cc$typ[sel] <- "b"    }    if (any(cc_tab %in% c("str", "e"))) {        sel <- cc$c_t %in% c("str", "e")        cc$val[sel] <- replaceXMLEntities(cc$v[sel])        cc$typ[sel] <- "s"    }    if (any(cc_tab %in% c("inlineStr"))) {        sel <- cc$c_t %in% c("inlineStr")        cc$val[sel] <- is_to_txt(cc$is[sel])        cc$typ[sel] <- "s"    }    if (any(cc_tab %in% c("s"))) {        sel <- cc$c_t %in% c("s")        cc$val[sel] <- sst[as.numeric(cc$v[sel]) + 1]        cc$typ[sel] <- "s"    }    has_na_string <- FALSE    if (!all(is.na(na.strings))) {        sel <- cc$val %in% na.strings        if (any(sel)) {            cc$val[sel] <- NA_character_            cc$typ[sel] <- "na_string"            has_na_string <- TRUE        }    }    has_na_number <- FALSE    na.numbers <- as.character(na.numbers)    if (!all(is.na(na.numbers))) {        sel <- cc$v %in% na.numbers        if (any(sel)) {            cc$val[sel] <- NA_character_            cc$typ[sel] <- "na_number"            has_na_number <- TRUE        }    }    origin <- get_date_origin(wb)    if (!is.null(cc$c_s)) {        if (detect_dates && missing(types)) {            cc$is_string <- FALSE            if (!is.null(cc$c_t))                 cc$is_string <- cc$c_t %in% c("s", "str", "b",                   "inlineStr")            if (any(sel <- cc$c_s %in% xlsx_date_style)) {                sel <- sel & !cc$is_string & cc$v != ""                cc$val[sel] <- suppressWarnings(as.character(convert_date(cc$v[sel],                   origin = origin)))                cc$typ[sel] <- "d"            }            if (any(sel <- cc$c_s %in% xlsx_hms_style)) {                sel <- sel & !cc$is_string & cc$v != ""                if (isNamespaceLoaded("hms")) {                  cc$val[sel] <- cc$v[sel]                }                else {                  cc$val[sel] <- suppressWarnings(as.character(convert_hms(cc$v[sel])))                }                cc$typ[sel] <- "h"            }            if (any(sel <- cc$c_s %in% xlsx_posix_style)) {                sel <- sel & !cc$is_string & cc$v != ""                cc$val[sel] <- suppressWarnings(as.character(convert_datetime(cc$v[sel],                   origin = origin)))                cc$typ[sel] <- "p"            }        }    }    if (any(sel <- is.na(cc$typ))) {        cc$val[sel] <- cc$v[sel]        cc$typ[sel] <- "n"    }    if (show_formula) {        if (any(cc$f_t == "shared")) {            cc_shared <- wb$worksheets[[sheet]]$sheet_data$cc            cc_shared <- cc_shared[cc_shared$f_t == "shared",                 ]            cc_shared <- cc_shared[order(as.integer(cc_shared$f_si)),                 ]            cc_shared$f <- ave2(cc_shared$f, cc_shared$f_si,                 carry_forward)            cc_shared$cols <- ave2(col2int(cc_shared$c_r), cc_shared$f_si,                 calc_distance)            cc_shared$rows <- ave2(as.integer(cc_shared$row_r),                 cc_shared$f_si, calc_distance)            cells <- find_a1_notation(cc_shared$f)            repls <- vector("list", length = length(cells))            for (i in seq_along(cells)) {                repls[[i]] <- next_cell(cells[[i]], cc_shared$cols[i],                   cc_shared$rows[i])            }            cc_shared$f <- replace_a1_notation(cc_shared$f, repls)            cc_shared$cols <- NULL            cc_shared$rows <- NULL            cc_shared <- cc_shared[which(cc_shared$r %in% cc$r),                 ]            cc[match(cc_shared$r, cc$r), ] <- cc_shared        }        sel <- cc$f != ""        cc$val[sel] <- replaceXMLEntities(cc$f[sel])        cc$typ[sel] <- "f"    }    if (show_hyperlinks) {        if (length(wb$worksheets[[sheet]]$hyperlinks)) {            hls <- wb_to_hyperlink(wb, sheet)            hyprlnks <- as.data.frame(do.call("rbind", lapply(hls,                 function(hl) {                  c(hl$ref, ifelse(is.null(hl$target), hl$location,                     hl$target))                })))            cc$val[match(hyprlnks$V1, cc$r)] <- hyprlnks$V2        }    }    if (has_na_string)         cc$typ[cc$typ == "na_string"] <- NA    if (has_na_number)         cc$typ[cc$typ == "na_number"] <- NA    zz <- cc[c("val", "typ")]    zz$cols <- NA_integer_    zz$rows <- NA_integer_    zz$cols <- match(cc$c_r, colnames(z)) - 1L    zz$rows <- match(cc$row_r, rownames(z)) - 1L    zz <- zz[order(zz[, "cols"], zz[, "rows"]), ]    if (any(zz$val == "", na.rm = TRUE))         zz <- zz[zz$val != "", ]    long_to_wide(z, tt, zz)    if (fill_merged_cells) {        mc <- wb$worksheets[[sheet]]$mergeCells        if (length(mc)) {            mc <- unlist(xml_attr(mc, "mergeCell"))            for (i in seq_along(mc)) {                filler <- stringi::stri_split_fixed(mc[i], pattern = ":")[[1]][1]                dms <- dims_to_dataframe(mc[i])                if (any(row_sel <- rownames(z) %in% rownames(dms)) &&                   any(col_sel <- colnames(z) %in% colnames(dms))) {                  z_fill <- wb_to_df(file = wb, sheet = sheet,                     dims = filler, na.strings = na.strings, convert = FALSE,                     col_names = FALSE, detect_dates = detect_dates,                     show_formula = show_formula, keep_attributes = TRUE)                  tt_fill <- attr(z_fill, "tt")                  z[row_sel, col_sel] <- z_fill                  tt[row_sel, col_sel] <- tt_fill                }            }        }    }    if (skip_hidden_rows) {        sel <- row_attr$hidden == "1" | row_attr$hidden == "true"        if (any(sel)) {            hide <- !keep_rows %in% row_attr$r[sel]            z <- z[hide, , drop = FALSE]            tt <- tt[hide, , drop = FALSE]        }    }    if (skip_hidden_cols) {        col_attr <- wb$worksheets[[sheet]]$unfold_cols()        sel <- col_attr$hidden == "1" | col_attr$hidden == "true"        if (any(sel)) {            hide <- col2int(keep_cols) %in% as.integer(col_attr$min[sel])            z[hide] <- NULL            tt[hide] <- NULL        }    }    if (skip_empty_rows) {        empty <- vapply(seq_len(nrow(z)), function(x) all(is.na(z[x,             ])), NA)        z <- z[!empty, , drop = FALSE]        tt <- tt[!empty, , drop = FALSE]    }    if (skip_empty_cols) {        empty <- vapply(z, function(x) all(is.na(x)), NA)        if (any(empty)) {            sel <- which(empty)            z[sel] <- NULL            tt[sel] <- NULL        }    }    xlsx_cols_names <- colnames(z)    names(xlsx_cols_names) <- xlsx_cols_names    if (col_names) {        nams <- names(xlsx_cols_names)        xlsx_cols_names <- z[1, ]        names(xlsx_cols_names) <- nams        z <- z[-1, , drop = FALSE]        tt <- tt[-1, , drop = FALSE]    }    if (row_names) {        rownames(z) <- z[, 1]        rownames(tt) <- z[, 1]        xlsx_cols_names <- xlsx_cols_names[-1]        z <- z[, -1, drop = FALSE]        tt <- tt[, -1, drop = FALSE]    }    date_conv <- NULL    datetime_conv <- NULL    hms_conv <- convert_hms    if (missing(types)) {        types <- guess_col_type(tt)        date_conv <- as.Date        datetime_conv <- as.POSIXct    }    else {        if (any(xlsx_cols_names %in% names(types))) {            if (is.character(types)) {                types[types == "character"] <- 0                types[types == "numeric"] <- 1                types[types == "Date"] <- 2                types[types == "POSIXct"] <- 3                types[types == "logical"] <- 4                types[types == "hms"] <- 5                types[types == "formula"] <- 6            }            if (any(!names(types) %in% xlsx_cols_names)) {                warning("variable from `types` not found in data")                types <- types[names(types) %in% xlsx_cols_names]            }            names(types) <- names(xlsx_cols_names[match(names(types),                 xlsx_cols_names)])            guess <- guess_col_type(tt)            guess[names(types)] <- types            types <- guess        }        else {            stop("no variable from `types` found in data")        }        date_conv <- convert_date        datetime_conv <- convert_datetime    }    if (convert) {        sel <- !is.na(names(types))        if (any(sel)) {            nums <- names(which(types[sel] == 1))            dtes <- names(which(types[sel] == 2))            poxs <- names(which(types[sel] == 3))            logs <- names(which(types[sel] == 4))            difs <- names(which(types[sel] == 5))            fmls <- names(which(types[sel] == 6))            if (length(nums))                 z[nums] <- lapply(z[nums], function(i) as.numeric(replace(i,                   i == "#NUM!", "NaN")))            if (length(dtes))                 z[dtes] <- lapply(z[dtes], date_conv, origin = origin)            if (length(poxs))                 z[poxs] <- lapply(z[poxs], datetime_conv, origin = origin)            if (length(logs))                 z[logs] <- lapply(z[logs], as.logical)            if (isNamespaceLoaded("hms"))                 z[difs] <- lapply(z[difs], hms_conv)            for (i in seq_along(z)) {                if (names(z)[i] %in% fmls)                   class(z[[i]]) <- c(class(z[[i]]), "formula")            }        }        else {            warning("could not convert. All missing in row used for variable names")        }    }    if (col_names) {        if (check_names) {            xlsx_cols_names <- make.names(xlsx_cols_names, unique = TRUE)        }        names(z) <- xlsx_cols_names        names(tt) <- xlsx_cols_names    }    if (keep_attributes) {        attr(z, "tt") <- tt        attr(z, "types") <- types        if (!missing(named_region))             attr(z, "dn") <- nr    }    z})(wb = <environment>, sheet = "Sheet 1"): argument "file" is missing, with no default